Re-apply phase 1 code
We needed to re-apply the phase 1 code, changing the tx fixture's
addresses.
