+++ Cargo.toml
@@ -8,7 +8,10 @@ edition = "2018"
 
 [dependencies]
 bytes = "1.0.1"
+cosmos-sdk-proto = "0.5.0"
+cosmos_sdk = "0.1.1"
 prost = "0.7.0"
+prost-types = "0.7.0"
 serde = "1.0"
 serde_json = "1.0"
 structopt = "0.3.21"
