+++ src/tx.rs
@@ -1,0 +1,10 @@
+//! Transaction modeling for the basecoin app.
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Transaction {
+    pub sender: String,
+    pub receiver: String,
+    pub amount: u64,
+}
