+++ tests/fixtures/tx.json
@@ -1,0 +1,31 @@
+{
+  "body": {
+    "messages": [
+      {
+        "@type": "/cosmos.bank.v1beta1.MsgSend",
+        "from_address": "cosmos1snd5m4h0wt5ur55d47vpxla389r2xkf8dl6g9w",
+        "to_address": "cosmos1t2e0nyjhwn3revunvf2uperhftvhzu4euuzva9",
+        "amount": [
+          {
+            "denom": "uatom",
+            "amount": "100"
+          }
+        ]
+      }
+    ],
+    "memo": "",
+    "timeout_height": "0",
+    "extension_options": [],
+    "non_critical_extension_options": []
+  },
+  "auth_info": {
+    "signer_infos": [],
+    "fee": {
+      "amount": [],
+      "gas_limit": "200000",
+      "payer": "",
+      "granter": ""
+    }
+  },
+  "signatures": []
+}
